% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxjson.R
\name{boxAtoms}
\alias{boxAtoms}
\title{Box atoms in JSON}
\usage{
boxAtoms(json, strict = TRUE)
}
\arguments{
\item{json}{JSON string or file reference.}

\item{strict}{Throw an error if input is not valid JSON?}
}
\value{
JSON string.
}
\description{
Box atoms in JSON
}

