% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\docType{package}
\name{boxjson}
\alias{boxjson}
\alias{boxjson-package}
\title{bounds: Box JSON atoms}
\description{
Utility to transform JSON from unboxed to boxed state and vice versa.
}
\section{Exports}{

  \code{\link{hasUnboxedAtom}}
  \code{\link{boxAtoms}}
  \code{\link{unboxAtoms}}
}

\section{Repository}{

  \url{https://github.com/chiefBiiko/boxjson}
}

