% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxjson.R
\name{hasUnboxedAtom}
\alias{hasUnboxedAtom}
\title{Check if JSON contains unboxed atoms}
\usage{
hasUnboxedAtom(json)
}
\arguments{
\item{json}{JSON string or file reference.}
}
\value{
Logical.
}
\description{
Check if JSON contains unboxed atoms
}

