% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{isStruct}
\alias{isStruct}
\title{Is JSON a (non-empty) array(-like) or object(-like)?}
\usage{
isStruct(json)
}
\arguments{
\item{json}{JSON string.}
}
\value{
Logical.
}
\description{
Is JSON a (non-empty) array(-like) or object(-like)?
}
\keyword{internal}
