% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unboxAtoms.R
\name{unboxAtoms}
\alias{unboxAtoms}
\title{Unbox atoms in JSON}
\usage{
unboxAtoms(json, strict = TRUE)
}
\arguments{
\item{json}{JSON string or file reference.}

\item{strict}{Throw an error if input is not valid JSON?}
}
\value{
JSON string.
}
\description{
Unbox atoms in JSON
}
